Final Tests stablished

+++ src/main/java/cat/itacademy/s05/t01/n01/S05T01N01/DTOs/responses/PlayerCreateResponseDTO.java
@@ -3,7 +3,7 @@ package cat.itacademy.s05.t01.n01.S05T01N01.DTOs.responses;
 public record PlayerCreateResponseDTO(
         long id,
         String name,
-        Integer balance,
+        Double balance,
         int gamesWon,
         int gamesPlayed
 ) {
